+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+### 2.1.1 - 2018-05-11
+
+- Subject can now be Symbols or any other object that responds top `#to_s`.
+
+- Fixed bug where publishes from the `#on_open` callback broke the connection.
+
 ### 2.1.0 - 2018-05-10
 
 - This is a minor release even though the API has changed. The changed API is the one for Rack based WebSocket and SSE connection upgrades. The PR for the spec addition is currently stalled but some suggestions for a stateless API are implemented in this release. The proposed Rack SPEC is [here](misc/SPEC). The PR is [here](https://github.com/rack/rack/pull/1272)